CategoryUPdatesDone

+++ app/Http/Controllers/Api/CategoryController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use App\Http\Resources\Category as CategoryResource;
+use App\Http\Resources\PopularCategory;
 use App\Http\Resources\Subcategory as ResourcesSubcategory;
 use App\Models\Category as CategoryModel;
 use App\Models\CategoryTrans;
@@ -72,6 +73,7 @@ class CategoryController extends Controller
                     $decoded_ct = json_decode($ct);
                     CategoryTrans::create([
                         'name' => $decoded_ct->name,
+                        'description' => $decoded_ct->description,
                         'locale' => $decoded_ct->locale,
                         'categ_id' => $categ->id
                     ]);
@@ -150,69 +152,56 @@ class CategoryController extends Controller
         try {
             // Validation of $id should goes here
 
-        
-           // dd($request->all());
+
+            $categ = CategoryModel::find($id);
+
+            if ($categ != null) {
+
+                // Validation of $request should goes here
+                ////
+
+                // add image to storage ////////////////////////////////
+                $catge_image = $request->image;
+                $new_categ_image = random_int(100000, 999999) . '.' . $catge_image->getClientOriginalExtension();
+                Storage::disk('public')->put($new_categ_image, file_get_contents($catge_image));
+
+                // Save to DB ///////////////////////////////////////////
+
+                $img_path = 'public/' . $categ->image;
+
+                Storage::delete($img_path);
+
+                $categ->image = $new_categ_image;
+                $categ->save();
+
+                if (is_array($request->categ_trans)) {
+                    foreach ($request->categ_trans as $key => $ct) {
+                        $decoded_ct = json_decode($ct);
+
+                        CategoryTrans::where('locale', $decoded_ct->locale)->where('categ_id', $id)->update(['name' => $decoded_ct->name, 'description' => $decoded_ct->description]);
+                    }
+
+                    return response()->json([
+                        'message' => "Object Updated"
+                    ], 200);
+
+                } else {
+                    return response()->json([
+                        'status' => true,
+                        'message' => 'categ_trans is not array'
+                    ], 500);
+                }
+            } else {
+                return response()->json([
+                    'message' => "Object Not Found"
+                ], 404);
+            }
 
             /////////////////////////////////////
 
-            // Validation of $request should goes here
-            ////
-
-            // add image to storage ////////////////////////////////
-            $catge_image = $request->image;
-            $new_categ_image = random_int(100000, 999999) . '.' . $catge_image->getClientOriginalExtension();
-            Storage::disk('public')->put($new_categ_image, file_get_contents($catge_image));
-
-            // Save to DB ///////////////////////////////////////////
-            $categ = CategoryModel::find($id);
-
-            $img_path = 'public/' . $categ->image;
-
-            Storage::delete($img_path);
-
-            $categ->image = $new_categ_image;
-            $categ->save();
-
-           
-            ///
-            // $categ_translation = null;
-            if (is_array($request->categ_trans)) {
-                foreach ($request->categ_trans as $key => $ct) {
-                    $decoded_ct = json_decode($ct);
-         
-                $categ_translation = CategoryTrans::where('locale',$decoded_ct->locale)->where('categ_id',$id)->update(['name'=>$decoded_ct->name]);
-                }
-                
-
-                return response()->json([
-                    'message' => "Done"],200);
-
-            } else {
-                return response()->json([
-                    'status' => true,
-                    'message' => 'categ_trans is not array'
-                ], 500);
-            }
-        
-
-         
 
 
-            // if (is_array($locals)) {
-            //     foreach ($locals as $key => $ct) {
-            //         dd($ct);
 
-            //     //  $locals[$key]=$decoded_ct->locale;
-            //     //     // CategoryTrans::create([
-            //     //     //     'name' => $decoded_ct->name,
-            //     //     //     'locale' => $decoded_ct->locale,
-            //     //     //     'categ_id' => $categ->id
-            //     //     // ]);
-            //      } 
-
-            //     }
-
-            
         } catch (\Throwable $th) {
             throw $th;
             // abort(code: 500, message: 'fail to update');
@@ -269,6 +258,53 @@ class CategoryController extends Controller
                 'status' => false,
                 'message' => $th->getMessage(),
             ], 500);
+        }
+    }
+    
+    public function popular(Request $request)
+    {
+        try {
+            foreach ($request->categ_ids as $key => $value) 
+            {
+                CategoryTrans::where('categ_id', $value)->update(['popular'=>1]);
+            }
+            return response()->json([
+                'message'=>'Object Updated'
+            ],200);
+        } catch (\Throwable $th) {
+            throw $th;
+        }
+    }
+
+    public function unpopular(Request $request)
+    {
+        try {
+            foreach ($request->categ_ids as $key => $value) 
+            {
+                CategoryTrans::where('categ_id', $value)->update(['popular'=>0]);
+            }
+            return response()->json([
+                'message'=>'Object Updated'
+            ],200);
+        } catch (\Throwable $th) {
+            throw $th;
+        }
+    }
+
+    public function getpopular()
+    {
+        try {
+            $popular_categ_query = CategoryModel::query();
+
+            $popular_categ_query->whereHas('categorytrans',function($query){
+                $query->where('popular',1);
+            });
+
+            $pop_categ = $popular_categ_query->get();
+
+            return PopularCategory::collection($pop_categ);
+        } catch (\Throwable $th) {
+            throw $th;
         }
     }
 }
